For #18900: Make Navigation Graph inflation asynchronous for performance purposes
For #18900: inflate navGraph programatically

For #169000: Made NavGraph inflation asynchronous

For #16900: Changed to block with runBlocking

For #16900: Refactored blocking call into a function

For 16900: NavGraph inflation is now async

We now attach the nav graph (or check if its attached) on every nav call ( an extension function for NavController).
This is done by checking the value of the job stored in PerfNavController.map which keeps track of the job with the NavController as a Key.
If the job hasn't been completed, it will block the main thread until the job is done. The job itself is responsible for attaching the navgraph
to the navcontroller (and the inflation of the latter too)

For 16900: Changed other calls of navigate to loadNavGraphBeforeNavigate

Some navigation methods are directly called from the NavController instead of the extension function .nav. Therefore, we need to make sure the
navgraph is loaded before we can call those navigation methods

For 16900: Rebase on master